requirements.txt
comtypes==1.1.7
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
Pillow==5.4.1
pyparsing==2.3.1
pypiwin32==223
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyserial==3.4
python-dateutil==2.8.0
PyVISA==1.9.1
pywin32==224
pywinauto==0.6.6
reportlab==3.5.13
six==1.12.0
SQLAlchemy==1.2.17
SQLAlchemy-Utils==0.33.11
xlrd==1.2.0
